stack.yaml: Add tasty-hunit dependency for tests
`stack build --test` (which runs a lot more than what .gh/wf/haskell
workflow does) takes a half hour or more and mostly passes, but the
test-nebula-plugin suite fails; it seems to be unable to spawn an
external program. (It's not `z3`; that's installed.) This can be
quickly reproduced with:

    stack build --test --ta '-p /add_assoc/'

From a quick look with strace, it appears that it may be trying to run
`cabal`, which of course doesn't exist in a pure Stack environment.
(Stack uses the cabal library, but not the command-line program.)

Nonetheless, this commit is still forward progress.